Edited Comment Service and ServiceImpl

+++ hunters-service/src/main/java/com/monsterhunters/pa165/service/CommentServiceImpl.java
@@ -3,6 +3,7 @@ package com.monsterhunters.pa165.service;
 import com.monsterhunters.pa165.dao.CommentDao;
 import com.monsterhunters.pa165.entity.Comment;
 import com.monsterhunters.pa165.entity.User;
+import com.monsterhunters.pa165.exceptions.HuntersServiceException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,7 +11,7 @@ import java.util.List;
 
 /**
  * Implementation of the {@link CommentService}.
- *
+ * <p>
  * Created by babcang on 17.11.2016.
  *
  * @author Babcan G
@@ -22,28 +23,56 @@ public class CommentServiceImpl implements CommentService {
     private CommentDao commentDao;
 
     @Override
-    public Comment createComment(Comment c) {
-        commentDao.create(c);
+    public Comment createComment(Comment c) throws HuntersServiceException {
+        if (c == null) {
+            throw new IllegalArgumentException("Comment is null");
+        }
+        try {
+            commentDao.create(c);
+        } catch (Throwable e) {
+            throw new HuntersServiceException("Something went wrong. " +
+                    "Comment " + c + " cannot be created", e);
+        }
         return c;
     }
 
     @Override
-    public void deleteComment(Comment c) {
-        commentDao.delete(c);
+    public void deleteComment(Comment c) throws HuntersServiceException {
+        if (c == null) {
+            throw new IllegalArgumentException("Weapon is null");
+        }
+        try {
+            commentDao.delete(c);
+        } catch (Throwable e) {
+            throw new HuntersServiceException("Comment " + c + " cannot be deleted", e);
+        }
+
     }
 
     @Override
-    public Comment findById(Long id) {
-        return commentDao.findById(id);
+    public Comment findById(Long id) throws HuntersServiceException {
+        try {
+            return commentDao.findById(id);
+        } catch (Throwable e) {
+            throw new HuntersServiceException("Cannot get comment with id:" + id, e);
+        }
     }
 
     @Override
-    public List<Comment> findAll() {
-        return commentDao.findAll();
+    public List<Comment> findAll() throws HuntersServiceException {
+        try {
+            return commentDao.findAll();
+        } catch (Throwable e) {
+            throw new HuntersServiceException("Cannot get list of comments", e);
+        }
     }
 
     @Override
-    public List<Comment> findByUser(User user) {
-        return commentDao.findByUser(user);
+    public List<Comment> findByUser(User user) throws HuntersServiceException {
+        try {
+            return commentDao.findByUser(user);
+        } catch (Throwable e) {
+            throw new HuntersServiceException("Cannot cooment of User: " + user.getNickname(), e);
+        }
     }
 }
